SI: Added the Data Download Button. This required to move all the data from ViewComponent into AppComponent. Controller and Model files has reference to ViewComponent changed it to App Component
