+++ src/app/app.component.css
@@ -4,3 +4,201 @@
   margin-top: 10px;
   margin-bottom: 10px;
 }
+.DownloadButton {
+  margin-right: 20px;
+  margin-left: 20px;
+  margin-top: 10px;
+  margin-bottom: 10px;
+}
+#Legend {
+  position: absolute;
+  display: block;
+  top: 50px;
+  right: 20px;
+  border: none;
+  font-size: 20pt;
+  z-index: 6;
+}
+
+#TimeseriesBox {
+  position: absolute;
+  bottom: 20px;
+  left: 20px;
+  z-index: 6;
+}
+
+#Legend text {
+  fill: white;
+  text-shadow: -1px -1px 0 #000000, 1px -1px 0 #000000, -1px 1px 0 #000000,
+    1px 1px 0 #000000, -2px 0 0 #000000, 2px 0 0 #000000, 0 2px 0 #000000,
+    0 -2px 0 #000000;
+}
+
+path.domain {
+  stroke: black;
+}
+
+.tick text {
+  fill: black;
+}
+
+.axis text {
+  font: 18px sans-serif;
+  fill: black;
+}
+
+.axis line,
+.axis path {
+  fill: none;
+  stroke: black;
+  shape-rendering: crispEdges;
+}
+
+.TabHeading {
+  text-align: center;
+}
+
+.mat-slider-horizontal {
+  min-width: 200px !important;
+  width: 200px;
+}
+
+.mat-ink-bar .mat-tab-label[_ngcontent-c3] {
+  width: 52px !important;
+}
+
+.mat-tab-label,
+.mat-tab-label-active {
+  min-width: 0 !important;
+  width: 60px !important;
+  padding: 4px !important;
+  margin: 4px !important;
+}
+
+#sideNavButton {
+  position: absolute;
+  left: 10px;
+  top: 10px;
+  z-index: 10;
+}
+
+/*.example-fab {
+  position: absolute;
+  right: 20px;
+  bottom: 10px;
+}*/
+
+#ClimateTitle {
+  position: absolute;
+  top: 10px;
+  left: 80px;
+  border: none;
+  text-align: center;
+  font-size: 30px;
+  z-index: 6;
+  color: white;
+  text-shadow: -1px -1px 0 #000000, 1px -1px 0 #000000, -1px 1px 0 #000000,
+    1px 1px 0 #000000, -2px 0 0 #000000, 2px 0 0 #000000, 0 2px 0 #000000,
+    0 -2px 0 #000000;
+}
+
+#ClimateDate {
+  position: absolute;
+  display: block;
+  bottom: 10px;
+  right: 20px;
+  border: none;
+  font-size: 20pt;
+  color: white;
+  text-shadow: /* Outline */ -1px -1px 0 #000000, 1px -1px 0 #000000,
+    -1px 1px 0 #000000, 1px 1px 0 #000000, -2px 0 0 #000000, 2px 0 0 #000000,
+    0 2px 0 #000000, 0 -2px 0 #000000;
+  z-index: 6;
+}
+
+html,
+body,
+material-app,
+md-sidenav-container,
+.my-content {
+  margin: 0;
+  width: 100%;
+  height: 100%;
+}
+
+.example-sidenav-fab-container .mat-sidenav-content,
+.example-sidenav-fab-container md-sidenav {
+  display: flex;
+  overflow: visible;
+}
+
+.radio-group {
+  display: inline-flex;
+  flex-direction: column;
+}
+
+.radio-button {
+  margin: 5px;
+}
+
+.InfoText {
+  max-width: 250px;
+  text-indent: 3ch;
+  margin: 10px;
+}
+
+.MatButton1 {
+  margin-right: 75px;
+  margin-left: 75px;
+  width: fit-content;
+  margin-bottom: 10px;
+}
+
+.MatButton2 {
+  margin-right: 50px;
+  margin-left: 50px;
+  width: fit-content;
+  margin-bottom: 10px;
+}
+
+.MatButton3 {
+  margin-right: 57px;
+  margin-left: 57px;
+  width: fit-content;
+  margin-bottom: 10px;
+}
+
+.MenTab {
+  font-size: 14px;
+}
+
+.mat-slide-toggle-content,
+.mat-radio-label-content {
+  font-size: 14px;
+  color: black;
+}
+
+.hidden {
+  visibility: hidden;
+  position: absolute;
+  left: 300px;
+  top: 50px;
+}
+
+.unhidden {
+  visibility: visible;
+}
+
+.close {
+  border: none;
+}
+
+button:focus {
+  outline: none;
+  border: none;
+  margin: 0px;
+}
+
+closebutton:focus {
+  background: none;
+}
